Incremental work on integrating the new layout editor UI.
This commit replaces the old (and functioning) editor with the new editor (using sample data).
Due to the nature of AngularJS and the fact that it loads templates dynamically, the iframe's height isn't properly adjusted.
We need to capture some client event whenever a template was loaded (and applied) to trigger an iframe resize.