Getting rid of the layout editor iframe.
This commit removes the layout editor iframe and renders it 'inline' instead.
The original reason for the iframe was to prevent CSS bleeding when rendering the front-end theme.
Later we decided to only render the admin theme, so we have full control over CSS, removing the need for the iframe.
This simplifies the implementation and prevents iframe resize issues.